update@rename embedding
